Rename test classes to avoid confusion
