+++ spring-test/src/test/java/org/springframework/test/web/client/match/XpathRequestMatchersTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2012 the original author or authors.
+ * Copyright 2002-2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.springframework.test.web.client.match;
 
 import java.io.IOException;
@@ -24,7 +25,7 @@ import org.junit.Test;
 import org.springframework.mock.http.client.MockClientHttpRequest;
 
 /**
- * Tests for {@link XpathRequestMatchers}.
+ * Unit tests for {@link XpathRequestMatchers}.
  *
  * @author Rossen Stoyanchev
  */
