+++ spring-test/src/test/java/org/springframework/test/web/client/samples/matchers/XmlContentRequestMatchersIntegrationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2012 the original author or authors.
+ * Copyright 2002-2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.springframework.test.web.client.samples.matchers;
 
 import java.net.URI;
@@ -42,11 +43,10 @@ import static org.springframework.test.web.client.response.MockRestResponseCreat
  * Examples of defining expectations on XML request content with XMLUnit.
  *
  * @author Rossen Stoyanchev
- *
- * @see ContentRequestMatcherTests
- * @see XpathRequestMatcherTests
+ * @see ContentRequestMatchersIntegrationTests
+ * @see XpathRequestMatchersIntegrationTests
  */
-public class XmlContentRequestMatcherTests {
+public class XmlContentRequestMatchersIntegrationTests {
 
 	private static final String PEOPLE_XML =
 			"<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>" +
